refactor(management): co-locate dto schemas and mappers
