+++ packages/management/src/dtos/index.ts
@@ -1,9 +1,26 @@
 export {
+	BulkActionResultSchema,
+	DeleteJobSchema,
+	ErrorSchema,
 	toBulkActionResultDto,
 	toDeleteJobDto,
+} from './actions.js';
+export { CapabilitiesSchema } from './capabilities.js';
+export {
+	JobCursorPageSchema,
+	JobSchema,
+	JobSelectorSchema,
+	RescheduleJobRequestSchema,
 	toJobCursorPageDto,
 	toJobDto,
+} from './job.js';
+export {
+	QueueStatsSchema,
+	QueueViewSummaryListSchema,
 	toQueueStatsDto,
 	toQueueViewSummaryListDto,
+} from './queue-view.js';
+export {
+	SchedulerHealthSchema,
 	toSchedulerHealthDto,
-} from './mapper.js';
+} from './scheduler-health.js';
