+++ packages/management/src/dtos/job.ts
@@ -1,28 +1,7 @@
+import type { CursorPage, PersistedJob } from '@monque/core';
 import { type Static, Type } from '@sinclair/typebox';
 
-export const SchedulerHealthSchema = Type.Object(
-	{
-		status: Type.Union([Type.Literal('ok'), Type.Literal('unavailable')]),
-		scheduler: Type.Object({
-			healthy: Type.Boolean(),
-		}),
-	},
-	{ $id: 'SchedulerHealth' },
-);
-
-export const CapabilitiesSchema = Type.Object(
-	{
-		readOnly: Type.Boolean(),
-		actions: Type.Object({
-			read: Type.Boolean(),
-			cancel: Type.Boolean(),
-			retry: Type.Boolean(),
-			reschedule: Type.Boolean(),
-			delete: Type.Boolean(),
-		}),
-	},
-	{ $id: 'Capabilities' },
-);
+import type { JobCursorPageDto, JobDto, ManagementOptions } from '../surface/index.js';
 
 const JobStatusSchema = Type.Union([
 	Type.Literal('pending'),
@@ -44,55 +23,6 @@ export const JobSelectorSchema = Type.Object(
 	{ $id: 'JobSelector', additionalProperties: false },
 );
 
-export const BulkActionResultSchema = Type.Object(
-	{
-		count: Type.Number(),
-		errors: Type.Array(
-			Type.Object({
-				jobId: Type.String(),
-				error: Type.String(),
-			}),
-		),
-	},
-	{ $id: 'BulkActionResult' },
-);
-
-export const QueueStatsSchema = Type.Object(
-	{
-		pending: Type.Number(),
-		processing: Type.Number(),
-		completed: Type.Number(),
-		failed: Type.Number(),
-		cancelled: Type.Number(),
-		total: Type.Number(),
-		avgProcessingDurationMs: Type.Optional(Type.Number()),
-	},
-	{ $id: 'QueueStats' },
-);
-
-export const QueueViewSummaryListSchema = Type.Object(
-	{
-		queueViews: Type.Array(
-			Type.Object({
-				name: Type.String(),
-				hasPersistedJobs: Type.Boolean(),
-				hasRegisteredWorker: Type.Boolean(),
-				stats: Type.Unsafe<Static<typeof QueueStatsSchema>>(
-					Type.Ref('#/components/schemas/QueueStats'),
-				),
-				worker: Type.Union([
-					Type.Object({
-						concurrency: Type.Number(),
-						activeCount: Type.Number(),
-					}),
-					Type.Null(),
-				]),
-			}),
-		),
-	},
-	{ $id: 'QueueViewSummaryList' },
-);
-
 export const JobSchema = Type.Object(
 	{
 		id: Type.String(),
@@ -111,7 +41,7 @@ export const JobSchema = Type.Object(
 		createdAt: Type.String({ format: 'date-time' }),
 		updatedAt: Type.String({ format: 'date-time' }),
 	},
-	{ $id: 'Job' },
+	{ $id: 'Job', additionalProperties: false },
 );
 
 export const JobCursorPageSchema = Type.Object(
@@ -121,14 +51,7 @@ export const JobCursorPageSchema = Type.Object(
 		hasNextPage: Type.Boolean(),
 		hasPreviousPage: Type.Boolean(),
 	},
-	{ $id: 'JobCursorPage' },
-);
-
-export const DeleteJobSchema = Type.Object(
-	{
-		deleted: Type.Literal(true),
-	},
-	{ $id: 'DeleteJob' },
+	{ $id: 'JobCursorPage', additionalProperties: false },
 );
 
 export const RescheduleJobRequestSchema = Type.Object(
@@ -138,9 +61,57 @@ export const RescheduleJobRequestSchema = Type.Object(
 	{ $id: 'RescheduleJobRequest', additionalProperties: false },
 );
 
-export const ErrorSchema = Type.Object(
-	{
-		error: Type.String(),
-	},
-	{ $id: 'ManagementError' },
-);
+export async function toJobCursorPageDto<TContext>(
+	options: ManagementOptions<TContext>,
+	page: CursorPage,
+	context: TContext,
+): Promise<JobCursorPageDto> {
+	return {
+		jobs: await Promise.all(
+			page.jobs.map((job) => toJobDto(options, job as PersistedJob, context)),
+		),
+		cursor: page.cursor,
+		hasNextPage: page.hasNextPage,
+		hasPreviousPage: page.hasPreviousPage,
+	};
+}
+
+export async function toJobDto<TContext>(
+	options: ManagementOptions<TContext>,
+	job: PersistedJob,
+	context: TContext,
+): Promise<JobDto> {
+	const serializePayload =
+		options.serializePayloadByJobName?.[job.name] ?? options.serializePayload;
+	const payload = serializePayload
+		? await serializePayload({ job, payload: job.data, context })
+		: job.data;
+	const dto: JobDto = {
+		id: job._id.toHexString(),
+		name: job.name,
+		status: job.status,
+		payload,
+		nextRunAt: job.nextRunAt.toISOString(),
+		lockedAt: job.lockedAt ? job.lockedAt.toISOString() : null,
+		claimedBy: job.claimedBy ?? null,
+		lastHeartbeat: job.lastHeartbeat ? job.lastHeartbeat.toISOString() : null,
+		failCount: job.failCount,
+		failureReason: job.failReason ?? null,
+		createdAt: job.createdAt.toISOString(),
+		updatedAt: job.updatedAt.toISOString(),
+	};
+
+	if (job.heartbeatInterval !== undefined) {
+		dto.heartbeatInterval = job.heartbeatInterval;
+	}
+
+	if (job.repeatInterval !== undefined) {
+		dto.repeatInterval = job.repeatInterval;
+	}
+
+	if (job.uniqueKey !== undefined) {
+		dto.uniqueKey = job.uniqueKey;
+	}
+
+	return dto;
+}
