--- packages/management/src/schemas/index.ts
@@ -1,13 +1,0 @@
-export {
-	BulkActionResultSchema,
-	CapabilitiesSchema,
-	DeleteJobSchema,
-	ErrorSchema,
-	JobCursorPageSchema,
-	JobSchema,
-	JobSelectorSchema,
-	QueueStatsSchema,
-	QueueViewSummaryListSchema,
-	RescheduleJobRequestSchema,
-	SchedulerHealthSchema,
-} from './dtos.js';
